fixed compiler warnings
git-svn-id: https://lmms.svn.sf.net/svnroot/lmms/trunk/lmms@1704 0778d3d1-df1d-0410-868b-ea421aaaa00d

+++ plugins/bit_invader/bit_invader.cpp
@@ -221,7 +221,7 @@ void bitInvader::loadSettings( const QDomElement & _this )
 
 void bitInvader::lengthChanged( void )
 {
-	m_graph.setLength( m_sampleLength.value() );
+	m_graph.setLength( (int) m_sampleLength.value() );
 
 	normalize();
 }
